Removing pointless role utils
