--- src/shared/utils/roles/am-admin.ts
@@ -1,7 +1,0 @@
-import { UserService } from "../../services";
-
-export default function amAdmin(
-  myUserInfo = UserService.Instance.myUserInfo,
-): boolean {
-  return myUserInfo?.local_user_view.local_user.admin ?? false;
-}
