--- src/shared/utils/roles/am-top-mod.ts
@@ -1,9 +1,0 @@
-import { CommunityModeratorView } from "lemmy-js-client";
-import { UserService } from "../../services";
-
-export default function amTopMod(
-  mods: CommunityModeratorView[],
-  myUserInfo = UserService.Instance.myUserInfo,
-): boolean {
-  return mods.at(0)?.moderator.id === myUserInfo?.local_user_view.person.id;
-}
